dat cleanup + optimization

+++ crates/holtburger-dat/benches/provider_bench.rs
@@ -1,4 +1,4 @@
-use criterion::{black_box, criterion_group, criterion_main, Criterion};
+use criterion::{Criterion, black_box, criterion_group, criterion_main};
 use holtburger_dat::{DatDatabase, DatFileType, HbaReader, HbaWriter, ResourceProvider};
 use std::path::PathBuf;
 use tempfile::NamedTempFile;
@@ -6,11 +6,15 @@ use tempfile::NamedTempFile;
 fn get_portal_dat_path() -> Option<PathBuf> {
     if let Ok(path_str) = std::env::var("HOLTBURGER_PORTAL_DAT") {
         let path = PathBuf::from(path_str);
-        if path.exists() { return Some(path); }
+        if path.exists() {
+            return Some(path);
+        }
     }
     let workspace_root = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../../");
     let fallback = workspace_root.join("ace-root/dats/client_portal.dat");
-    if fallback.exists() { return Some(fallback); }
+    if fallback.exists() {
+        return Some(fallback);
+    }
     None
 }
 
@@ -21,9 +25,11 @@ fn bench_providers(c: &mut Criterion) {
     };
 
     let dat_db = DatDatabase::new(&dat_path).expect("Failed to open DAT");
-    
+
     // Select 50 file IDs to cycle through
-    let ids: Vec<u32> = dat_db.files.keys()
+    let ids: Vec<u32> = dat_db
+        .files
+        .keys()
         .filter(|&&id| DatFileType::from_id(id).is_essential())
         .take(50)
         .cloned()
@@ -34,9 +40,9 @@ fn bench_providers(c: &mut Criterion) {
     writer.set_compression(true);
     for &id in &ids {
         let data = dat_db.get_file(id).unwrap();
-        writer.add(id, (id >> 24) as u32, data);
+        writer.add(id, id >> 24, data);
     }
-    
+
     let temp_hba = NamedTempFile::new().unwrap();
     writer.write(temp_hba.path()).unwrap();
     let hba = HbaReader::open(temp_hba.path()).unwrap();
